chore: created javadoc and refactored project
