+++ feign-retry-annotation/src/main/java/io/github/maurigre/cloud/common/feign/FeignRetryAutoConfiguration.java
@@ -1,6 +1,6 @@
-package io.github.mgr.cloud.common.feign;
+package io.github.maurigre.cloud.common.feign;
 
-import io.github.mgr.cloud.common.feign.config.FeignRetryAspect;
+import io.github.maurigre.cloud.common.feign.config.FeignRetryAspect;
 import org.springframework.boot.autoconfigure.condition.ConditionalOnClass;
 import org.springframework.context.EnvironmentAware;
 import org.springframework.context.annotation.Bean;
@@ -9,6 +9,10 @@ import org.springframework.context.annotation.EnableAspectJAutoProxy;
 import org.springframework.core.env.Environment;
 import org.springframework.retry.support.RetryTemplate;
 
+
+/**
+ * The type Feign retry auto configuration.
+ */
 @EnableAspectJAutoProxy
 @Configuration(proxyBeanMethods = false)
 @ConditionalOnClass(RetryTemplate.class)
@@ -16,6 +20,12 @@ public class FeignRetryAutoConfiguration implements EnvironmentAware {
 
     private Environment environment;
 
+
+    /**
+     * Feign retry aspect feign retry aspect.
+     *
+     * @return the feign retry aspect
+     */
     @Bean
 	public FeignRetryAspect feignRetryAspect() {
 		return new FeignRetryAspect(this.environment);
