+++ feign-retry-annotation/src/main/java/io/github/maurigre/cloud/common/feign/annotations/EnableFeignRetry.java
@@ -1,6 +1,6 @@
-package io.github.mgr.cloud.common.feign.config;
+package io.github.maurigre.cloud.common.feign.annotations;
 
-import io.github.mgr.cloud.common.feign.FeignRetryAutoConfiguration;
+import io.github.maurigre.cloud.common.feign.FeignRetryAutoConfiguration;
 import org.springframework.context.annotation.Import;
 
 import java.lang.annotation.ElementType;
@@ -9,6 +9,9 @@ import java.lang.annotation.RetentionPolicy;
 import java.lang.annotation.Target;
 
 
+/**
+ * The interface Enable feign retry.
+ */
 @Target(ElementType.TYPE)
 @Retention(RetentionPolicy.RUNTIME)
 @Import(FeignRetryAutoConfiguration.class)
