--- feign-retry-annotation/src/main/java/io/github/mgr/cloud/common/feign/annotations/Backoff.java
@@ -1,21 +1,0 @@
-package io.github.mgr.cloud.common.feign.annotations;
-
-import java.lang.annotation.Documented;
-import java.lang.annotation.ElementType;
-import java.lang.annotation.Inherited;
-import java.lang.annotation.Retention;
-import java.lang.annotation.RetentionPolicy;
-import java.lang.annotation.Target;
-
-@Target(ElementType.TYPE)
-@Retention(RetentionPolicy.RUNTIME)
-@Documented
-@Inherited
-public @interface Backoff {
-
-  String delay() default "1000";
-
-  String maxDelay() default "0";
-
-  String multiplier() default "0.0";
-}
