+++ pubsub-annotation/src/main/java/io/github/maurigre/cloud/common/pubsub/PubsubAutoConfiguration.java
@@ -1,13 +1,17 @@
-package io.github.mgr.cloud.common.pubsub;
+package io.github.maurigre.cloud.common.pubsub;
 
-import io.github.mgr.cloud.common.pubsub.config.PubSubListenerRegister;
-import io.github.mgr.cloud.common.pubsub.config.PubSubProperties;
+import io.github.maurigre.cloud.common.pubsub.config.PubSubListenerRegister;
+import io.github.maurigre.cloud.common.pubsub.config.PubSubProperties;
 import org.springframework.boot.context.properties.EnableConfigurationProperties;
 import org.springframework.context.ApplicationContext;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
 import org.springframework.core.env.Environment;
 
+
+/**
+ * The type Pubsub auto configuration.
+ */
 @Configuration
 @EnableConfigurationProperties(PubSubProperties.class)
 public class PubsubAutoConfiguration {
@@ -16,13 +20,25 @@ public class PubsubAutoConfiguration {
     private final PubSubProperties pubSubProperties;
     private final Environment environment;
 
-    public PubsubAutoConfiguration(ApplicationContext applicationContext, PubSubProperties pubSubProperties, Environment environment) {
+  /**
+   * Instantiates a new Pubsub auto configuration.
+   *
+   * @param applicationContext the application context
+   * @param pubSubProperties   the pub sub properties
+   * @param environment        the environment
+   */
+  public PubsubAutoConfiguration(ApplicationContext applicationContext, PubSubProperties pubSubProperties, Environment environment) {
         this.applicationContext = applicationContext;
         this.pubSubProperties = pubSubProperties;
         this.environment = environment;
     }
 
-    @Bean
+  /**
+   * Pub sub listener register pub sub listener register.
+   *
+   * @return the pub sub listener register
+   */
+  @Bean
     public PubSubListenerRegister pubSubListenerRegister() {
         return new PubSubListenerRegister(applicationContext, pubSubProperties, environment);
     }
