+++ pubsub-annotation/src/main/java/io/github/maurigre/cloud/common/pubsub/annotations/EnablePubsub.java
@@ -1,6 +1,6 @@
-package io.github.mgr.cloud.common.pubsub.config;
+package io.github.maurigre.cloud.common.pubsub.annotations;
 
-import io.github.mgr.cloud.common.pubsub.PubsubAutoConfiguration;
+import io.github.maurigre.cloud.common.pubsub.PubsubAutoConfiguration;
 import java.lang.annotation.ElementType;
 import java.lang.annotation.Retention;
 import java.lang.annotation.RetentionPolicy;
@@ -8,6 +8,9 @@ import java.lang.annotation.Target;
 import org.springframework.context.annotation.Import;
 
 
+/**
+ * The interface Enable pubsub.
+ */
 @Target(ElementType.TYPE)
 @Retention(RetentionPolicy.RUNTIME)
 @Import(PubsubAutoConfiguration.class)
