+++ pubsub-annotation/src/main/java/io/github/maurigre/cloud/common/pubsub/annotations/PubSubListener.java
@@ -1,4 +1,4 @@
-package io.github.mgr.cloud.common.pubsub.annotations;
+package io.github.maurigre.cloud.common.pubsub.annotations;
 
 import java.lang.annotation.Documented;
 import java.lang.annotation.ElementType;
@@ -8,11 +8,25 @@ import java.lang.annotation.RetentionPolicy;
 import java.lang.annotation.Target;
 import org.springframework.stereotype.Component;
 
+/**
+ * The interface Pub sub listener.
+ */
 @Target({ElementType.METHOD, ElementType.TYPE})
 @Retention(RetentionPolicy.RUNTIME)
 @Documented
 @Inherited
 public @interface PubSubListener {
-    String subscription();
-    String errorHandler();
+  /**
+   * Subscription string.
+   *
+   * @return the string
+   */
+  String subscription();
+
+  /**
+   * Error handler string.
+   *
+   * @return the string
+   */
+  String errorHandler();
 }
