+++ pubsub-annotation/src/main/java/io/github/maurigre/cloud/common/pubsub/config/PubSubListenerRegister.java
@@ -1,7 +1,7 @@
-package io.github.mgr.cloud.common.pubsub.config;
+package io.github.maurigre.cloud.common.pubsub.config;
 
-import io.github.mgr.cloud.common.pubsub.annotations.PubSubListener;
-import io.github.mgr.cloud.common.pubsub.handlers.PubSubListenerErrorHandler;
+import io.github.maurigre.cloud.common.pubsub.annotations.PubSubListener;
+import io.github.maurigre.cloud.common.pubsub.handlers.PubSubListenerErrorHandler;
 import com.google.cloud.pubsub.v1.AckReplyConsumer;
 import com.google.cloud.pubsub.v1.MessageReceiver;
 import com.google.cloud.pubsub.v1.Subscriber;
@@ -14,6 +14,7 @@ import java.util.List;
 import java.util.Map;
 import java.util.concurrent.atomic.AtomicReference;
 import java.util.stream.Collectors;
+import lombok.NoArgsConstructor;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.apache.commons.lang3.StringUtils;
@@ -24,6 +25,12 @@ import org.springframework.context.ApplicationContext;
 import org.springframework.core.env.Environment;
 import org.springframework.stereotype.Component;
 
+/**
+ * Responsible for registering Pub/Sub listeners during application initialization.
+ *
+ * <p>This class uses a constructor generated by Lombok's {@code @RequiredArgsConstructor}
+ * annotation to initialize the required dependencies.
+ */
 @Slf4j
 @Component
 @RequiredArgsConstructor
@@ -35,9 +42,6 @@ public class PubSubListenerRegister implements SmartInitializingSingleton {
     private final Environment environment;
     private final List<Subscriber> subscribers = new ArrayList<>();
     private final AtomicReference<String> basePackageAtomicReference = new AtomicReference<>();
-
-
-
 
     @Override
     public void afterSingletonsInstantiated() {
@@ -108,7 +112,10 @@ public class PubSubListenerRegister implements SmartInitializingSingleton {
         }
     }
 
-    @PreDestroy
+  /**
+   * On shutdown void.
+   */
+  @PreDestroy
     public void onShutdown() {
         for (Subscriber subscriber : subscribers) {
             if (subscriber != null) {
@@ -166,7 +173,12 @@ public class PubSubListenerRegister implements SmartInitializingSingleton {
         return true;
     }
 
-    public String getBasePackage() {
+  /**
+   * Gets base package.
+   *
+   * @return the base package
+   */
+  public String getBasePackage() {
         return basePackageAtomicReference.updateAndGet(currentValue -> {
             if (currentValue == null) {
                 Map<String, Object> beansWithAnnotation = applicationContext.getBeansWithAnnotation(SpringBootApplication.class);
